requirements.txt
streamlit
llama-index
nltk
watchdog
python-dotenv